Refactored Nonprofits page and AccordionElement to correct logic and simplify code structure; Renamed AccordionElement to OrganizationAccordion
